pyproject.toml
[project]
name = "mymqtt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "paho-mqtt>=2.0.0",
    "streamlit>=1.54.0",
]

[project.scripts]
mymqtt = "main:main"
